+++ src/Settings/PortConfigurationView.tsx
@@ -1,13 +1,11 @@
 import {
   Button,
-  Checkbox,
   Select,
   MenuItem,
   FormControl,
   InputLabel,
   Box,
   ButtonPropsColorOverrides,
-  FormControlLabel,
   Typography,
 } from '@mui/material';
 import { OverridableStringUnion } from '@mui/types';