+++ CHANGELOG.md
@@ -9,11 +9,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
-- Added ability to change terminal font size.
+- Added ability to change terminal font size to both settings menu and terminal toolbar.
 
 ### Changed
 
 - Combined view and model files together into a single directory tree.
+- Setting inputs can now apply on loss of focus or enter rather than a Submit button.
 
 ### Fixed